# feat: Speed up tests on Apple Silicon and CI