M1 Macs don't have SMT, so there's no need to halve the processors when running tests

Before:
BUILD SUCCESSFUL in 3m 1s
BUILD SUCCESSFUL in 2m 57s

After:
BUILD SUCCESSFUL in 2m 08s
BUILD SUCCESSFUL in 1m 53s

Note: Scenario was unit test runs inside Android Studio, had similar speedup with `./gradlew clean jacocoUnitTestReport --rerun-tasks`

## How Has This Been Tested?

All tests
## Learning (optional, can help others)

We may want to do this based on physical CPUs if there is a good platform-agnostic method.

```
davidallison@Davids-MBP optilingo_mobile_app_v2 % sysctl -n hw.physicalcpu                                         
10
davidallison@Davids-MBP optilingo_mobile_app_v2 % sysctl -n hw.ncpu                                 
10
```


## Checklist
_Please, go through these checks before submitting the PR._

- [x] You have not changed whitespace unnecessarily (it makes diffs hard to read)
- [x] You have a descriptive commit message with a short title (first line, max 50 chars).
- [x] Your code follows the style of the project (e.g. never omit braces in `if` statements) 
- [x] You have commented your code, particularly in hard-to-understand areas
- [x] You have performed a self-review of your own code
- [ ] UI changes: include screenshots of all affected screens (in particular showing any new or changed strings)
- [ ] UI Changes: You have tested your change using the [Google Accessibility Scanner](https://play.google.com/store/apps/details?id=com.google.android.apps.accessibility.auditor)