@mikehardy 2022-05-13
Nice! For reference on intel mac, it generates what seems to work for you stated possible other version

```
mike@osxvm:~ % sysctl -n hw.physicalcpu
6
mike@osxvm:~ % sysctl -n hw.ncpu 
12
```

Additionally, there is perhaps an opportunity for CI tuning "while we're here" - that is, I have this information from github's doc:

```
Hardware specification for Windows and Linux virtual machines:

2-core CPU
7 GB of RAM memory
14 GB of SSD disk space
Hardware specification for macOS virtual machines:

3-core CPU
14 GB of RAM memory
14 GB of SSD disk space
```

So it could be that 2 is the right number for CI in general, and 3 is the right number for macOS in CI ? Then hw.physicalcpu if it's !CI && macOS-arm64 ?