@mikehardy 2022-05-13
using hardware cpu count would also mean it works even if people are using x86-64 JVM still, on arm64 hardware