@mikehardy 2022-05-13
An interesting point, the problem of course is that they are virtual machines, so the physical topology presented to the virtual machine and reported in /proc/cpuinfo is fictitious

The only way to know is to bench it both ways :laughing: