@mikehardy 2022-05-13
(by the way, my intel mac cpu info above is also fictitious, that was from a VM, but I happen to know that I presented an imaginary CPU topology that was SMT with 2 threads per core, so the experimental result is valid - I don't know what hardware resources they are actually presenting in azure cloud as their "vCPU", just an SMT thread or a real core)

It appears that for this instance type we are 1:1 so we should use 2 cores!

https://docs.microsoft.com/en-gb/azure/virtual-machines/acu

```


DS1_v2 - DS15_v2 | 210 - 250* | 1:1
-- | -- | --


[DS1_v2 - DS15_v2](https://docs.microsoft.com/en-gb/azure/virtual-machines/dv2-dsv2-series)	210 - 250*	1:1
```