@mikehardy 2022-05-13
Looking at the recent actions timings vs this ones, I think this is a really significant CI speedup, except windows. I booted my windows VM today and got it up to date, I'll see what I can do about memory optimization there since all indications are that's the issue